// Cast added on Shop ID (int)
git-svn-id: http://dev.prestashop.com/svn/v1/branches/1.5.x@16215 b9a71923-0436-4b27-9f14-aed3839534dd

+++ modules/blocktopmenu/menutoplinks.class.php
@@ -29,7 +29,7 @@ class MenuTopLinks
 {
 	public static function gets($id_lang, $id_linksmenutop = null, $id_shop)
 	{
-		$id_shop = ((bool) Configuration::get('PS_MULTISHOP_FEATURE_ACTIVE')) ? $id_shop : 0;
+		$id_shop = (int)((bool) Configuration::get('PS_MULTISHOP_FEATURE_ACTIVE')) ? $id_shop : 0;
 
 		$sql = 'SELECT l.id_linksmenutop, l.new_window, ll.link, ll.label
 				FROM '._DB_PREFIX_.'linksmenutop l
